Added kafka support

+++ src/main/java/com/scaler/orderservice/dto/OrderItemResponseDto.java
@@ -3,12 +3,14 @@ package com.scaler.orderservice.dto;
 import lombok.Builder;
 import lombok.Getter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Builder
 public class OrderItemResponseDto {
 
     private Long productId;
-    private Double price;
+    private BigDecimal price;
     private Integer quantity;
-    private Double itemTotal;
+    private BigDecimal itemTotal;
 }
