+++ src/main/java/com/scaler/orderservice/dto/OrderResponseDto.java
@@ -4,6 +4,7 @@ import com.scaler.orderservice.entity.OrderStatus;
 import lombok.Builder;
 import lombok.Getter;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.List;
 
@@ -13,12 +14,10 @@ public class OrderResponseDto {
 
     private Long id;
     private String userEmail;
-    private Double totalAmount;
+    private BigDecimal totalAmount;
     private OrderStatus status;
-
     private String paymentId;
     private LocalDateTime paidAt;
     private LocalDateTime createdAt;
-
     private List<OrderItemResponseDto> items;
 }
